Drop extraneous sed-ing of cma= in z50-raspi-firmware hook for Pi 4.
For starters, the pattern no longer exists since this commit in
raspi-firmware:

    commit dd456f4746a800ac85bdf376b5efcdb1fac133de
    Author: Gunnar Wolf <gwolf@gwolf.org>
    Date:   Wed Aug 5 12:02:57 2020 -0500

        Don't set CMA in RPi4 unless specified expressly

and there's now a SET_CMA variable instead.

And more importantly, the Pi 4 gets appropriate treatment thanks to this
commit (empty SET_CMA), which the Pi Compute Module 4 might get soon too
(see #996937).

This commit first shipped in debian/1.20200601-2, and we are using one
of those at the moment for the Pi 4 family:
 - 1.20210303+ds-2 (bullseye)
 - 1.20210303+ds-2~bpo10+1 (buster-backports)